refactor: use backend_api crate and reorganize code
