+++ Cargo.toml
@@ -27,6 +27,7 @@ serde = "1.0"
 uuid = "1.6"
 serde_bytes = "0.11"
 chrono = { version = "0.4", default-features = false, features = ["std"] }
+base64 = "0.22"
 
 mockall = "0.12"
 rstest = "0.18"
