+++ src/backend/impl/Cargo.toml
@@ -32,7 +32,7 @@ uuid = { workspace = true, features = ["serde"] }
 chrono.workspace = true
 hex = "0.4"
 lazy_static = "1.4"
-base64 = "0.22"
+base64.workspace = true
 
 rand = { version = "0.8", default-features = false }
 rand_chacha = { version = "0.3", default-features = false }
